add {baserelpath}
With this, you can set ARTICLE_SAVE_AS= or ARTICLE_URL= to
"{baserelpath}.html" and the generated filename will retain any folder
structure from the original .rst file. I needed this on my blog to
retain URL compatibility with an earlier system (which used
subdirectories to represent categories). This also removes the
possibility of accidental collisions in the output filename or URL,
since it's impossible to have "collisions" in the input pathname.